Added the crappy new tool: enumdirs from 3:2c3796e5d936
This crappy tool will generate a flat directory structure based on a
hierarchy of directories. It keeps the order between the files in each
subdirectory by adding a counter number in each one.

This way it is possible to generate a flat mp3 list to burn in a cd without
having musics from different artists/ambums being placed together.

Real crap. There must be a simpler solution.